updated all checks for service layers, and restructured service layer interface and class methods

+++ src/main/java/org/kylecodes/gm/services/ExerciseServiceImpl.java
@@ -1,17 +1,17 @@
 package org.kylecodes.gm.services;
 
+import org.kylecodes.gm.constants.RequestFailure;
 import org.kylecodes.gm.dtos.ExerciseDto;
 import org.kylecodes.gm.entities.Exercise;
 import org.kylecodes.gm.entities.User;
 import org.kylecodes.gm.entities.Workout;
 import org.kylecodes.gm.exceptions.ExerciseNotFoundException;
-import org.kylecodes.gm.constants.RequestFailure;
 import org.kylecodes.gm.exceptions.WorkoutNotFoundException;
-import org.kylecodes.gm.utils.SecurityUtil;
-import org.kylecodes.gm.repositories.ExerciseRepository;
-import org.kylecodes.gm.repositories.WorkoutRepository;
 import org.kylecodes.gm.mappers.EntityToDtoMapper;
 import org.kylecodes.gm.mappers.ExerciseToExerciseDtoMapper;
+import org.kylecodes.gm.repositories.ExerciseRepository;
+import org.kylecodes.gm.repositories.WorkoutRepository;
+import org.kylecodes.gm.utils.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,6 +26,7 @@ public class ExerciseServiceImpl implements ExerciseService{
     @Autowired
     private WorkoutRepository workoutRepository;
 
+
     EntityToDtoMapper<Exercise, ExerciseDto> exerciseMapper = new ExerciseToExerciseDtoMapper();
 
     @Override
@@ -38,25 +39,27 @@ public class ExerciseServiceImpl implements ExerciseService{
     @Override
     public List<ExerciseDto> getAllExercisesInWorkout(Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUserId(workoutId, user.getId())
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
-        List<Exercise> exercisesInWorkout = exerciseRepository.findAllByWorkout(workout.get());
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.GET_REQUEST_FAILURE);
+        }
+        List<Exercise> exercisesInWorkout = exerciseRepository.findAllByWorkoutId(workoutId);
         return exercisesInWorkout.stream().map((exercise) -> exerciseMapper.mapToDto(exercise)).toList();
     }
 
     @Override
-    public ExerciseDto getExerciseInWorkoutById(Long workoutId, Long exerciseId) {
+    public ExerciseDto getExerciseInWorkoutById(Long exerciseId, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUserId(workoutId, user.getId())
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
-        Exercise exercise = exerciseRepository.findByIdAndWorkout(exerciseId, workout.get()).orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.GET_REQUEST_FAILURE));
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.GET_REQUEST_FAILURE);
+        }
+        Exercise exercise = exerciseRepository.findByIdAndWorkoutId(exerciseId, workoutId).orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.GET_REQUEST_FAILURE));
         return exerciseMapper.mapToDto(exercise);
     }
 
     @Override
     public ExerciseDto createExercise(ExerciseDto exerciseDto, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.ofNullable(workoutRepository.findByIdAndUserId(workoutId, user.getId())
+        Optional<Workout> workout = Optional.of(workoutRepository.findByIdAndUserId(workoutId, user.getId())
                 .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
         Exercise exercise = new Exercise();
@@ -65,18 +68,17 @@ public class ExerciseServiceImpl implements ExerciseService{
 
         Exercise newExercise = exerciseRepository.save(exercise);
 
-        ExerciseDto exerciseResponse = exerciseMapper.mapToDto(newExercise);
-
-        return exerciseResponse;
+        return exerciseMapper.mapToDto(newExercise);
     }
 
     @Override
-    public ExerciseDto updateExerciseInWorkoutById(ExerciseDto exerciseDto, Long workoutId, Long exerciseId) {
+    public ExerciseDto updateExerciseInWorkoutById(ExerciseDto exerciseDto, Long exerciseId, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.of(workoutRepository.findByIdAndUserId(workoutId, user.getId())
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
-        Optional<Exercise> exercise = Optional.of(exerciseRepository.findByIdAndWorkout(exerciseId, workout.get())
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.GET_REQUEST_FAILURE);
+        }
+        Optional<Exercise> exercise = Optional.of(exerciseRepository.findByIdAndWorkoutId(exerciseId, workoutId)
                 .orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.PUT_REQUEST_FAILURE)));
 
         Exercise updateExercise = exercise.get();
@@ -89,13 +91,15 @@ public class ExerciseServiceImpl implements ExerciseService{
     }
 
     @Override
-    public void deleteExerciseInWorkoutById(Long workoutId, Long exerciseId) {
+    public void deleteExerciseInWorkoutById(Long exerciseId, Long workoutId) {
         User user = SecurityUtil.getPrincipalFromSecurityContext();
-        Optional<Workout> workout = Optional.of(workoutRepository.findByIdAndUserId(workoutId, user.getId())
-                .orElseThrow(() -> new WorkoutNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE)));
+        if (!workoutRepository.existsByIdAndUserId(workoutId, user.getId())) {
+            throw new WorkoutNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE);
+        }
 
-        Optional<Exercise> exercise = Optional.of(exerciseRepository.findByIdAndWorkout(exerciseId, workout.get()))
-                .orElseThrow(() -> new ExerciseNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE));
+        if (!exerciseRepository.existsByIdAndWorkoutId(exerciseId, workoutId)) {
+            throw new ExerciseNotFoundException(RequestFailure.DELETE_REQUEST_FAILURE);
+        }
         exerciseRepository.deleteById(exerciseId);
     }
